Allow for transactions that don't extend the file.
This changes the write pattern as well as the format of the header and
data. The header is now an array of 16 header entries, containing the
file offset, index within the group, and version of the data. The data
pages now contain a pattern that's made up of their index within the
group and their version.

The file size is increased by 16 pages after every 128 writes.
The 128 writes correspond to updating each of the 16 new pages 8 times.
Each write consists of writing a full page of data, followed by updating
the index on page 0 to reflect the newly-written data.